+++ lib/segment/analytics.rb
@@ -7,18 +7,18 @@ require 'segment/analytics/request'
 require 'segment/analytics/response'
 require 'segment/analytics/logging'
 
-module Segment
+module SegmentIo
   class Analytics
-    # Initializes a new instance of {Segment::Analytics::Client}, to which all
+    # Initializes a new instance of {SegmentIo::Analytics::Client}, to which all
     # method calls are proxied.
     #
     # @param options includes options that are passed down to
-    #   {Segment::Analytics::Client#initialize}
+    #   {SegmentIo::Analytics::Client#initialize}
     # @option options [Boolean] :stub (false) If true, requests don't hit the
     #   server and are stubbed to be successful.
     def initialize(options = {})
       Request.stub = options[:stub] if options.has_key?(:stub)
-      @client = Segment::Analytics::Client.new options
+      @client = SegmentIo::Analytics::Client.new options
     end
 
     def method_missing(message, *args, &block)