+++ lib/segment/analytics/request.rb
@@ -7,12 +7,12 @@ require 'net/http'
 require 'net/https'
 require 'json'
 
-module Segment
+module SegmentIo
   class Analytics
     class Request
-      include Segment::Analytics::Defaults::Request
-      include Segment::Analytics::Utils
-      include Segment::Analytics::Logging
+      include SegmentIo::Analytics::Defaults::Request
+      include SegmentIo::Analytics::Utils
+      include SegmentIo::Analytics::Logging
 
       # public: Creates a new request object to send analytics batch
       #
@@ -24,7 +24,7 @@ module Segment
         @path = options[:path] || PATH
         @retries = options[:retries] || RETRIES
         @backoff_policy =
-          options[:backoff_policy] || Segment::Analytics::BackoffPolicy.new
+          options[:backoff_policy] || SegmentIo::Analytics::BackoffPolicy.new
 
         http = Net::HTTP.new(options[:host], options[:port])
         http.use_ssl = options[:ssl]