+++ spec/segment/analytics/message_batch_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-module Segment
+module SegmentIo
   class Analytics
     describe MessageBatch do
       subject { described_class.new(100) }